Grace Access
This is basically just for making temporary backdoors.
It sends a request to a server and checks the response, differentiating between unavailable, wait, paid and unpaid and executing code accordingly.

This should not be used for any malicious purposes and only for very edgy case and for the sellers own safety.